+++ bin/wc-auto-refresh-all.sh
@@ -1,0 +1,6 @@
+#!/bin/sh
+while true
+do
+      sleep ${WC_SLEEP}
+      ./wc-refresh-all.sh 
+done